backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.35.2
torch==2.1.0
google-generativeai==0.3.1
python-dotenv==1.0.0
nltk==3.8.1
PyPDF2==3.0.1
python-multipart==0.0.6
pydantic==2.5.0